OSP12_create_container_registry.yaml - Removed the need to set include_ceph manually and instead, dynamically setting this fact based on the deployment vars.
create_templates.yaml - If deploying without Ceph and you include the resource_registry info for CephMon, CephOSD and CephClient, it still tries to configure
the controller nodes for these services which ultimately fail your deployment.  Added tasks to remove these lines from storage-environment.yaml if we are
not using Ceph.