Had a number of changes here.  Many were around re-structuring the playbook into something that made more sense. Moved dynamic template creation up in the process so that the resultant files could take advantage of the tasks which fixed the relative to absolute path names.
Also added in the network-environment.yaml as well as the network_data.yaml to the deployment script.  Testing
showed that if I included the network-environment.yaml and not the network_data.yaml, the fixed VIP's would
fail stating that they were not part of the appropriate subnet.  Including both fixed this, but I am not really
sure why it's throwing that error, so more investigation is probably needed there.